move getEventEmails to function
